+++ src/parser/part/prelude.rs
@@ -35,7 +35,7 @@ pub(crate) use crate::parser::lex::{
     starts_with_any_keyword, starts_with_keyword, ws1, ws_and_comments, PART_BODY_STARTERS,
 };
 pub(crate) use crate::parser::metadata::{metadata_def, metadata_usage};
-pub(crate) use crate::parser::metadata_annotation::{annotation, metadata_annotation};
+pub(crate) use crate::parser::metadata_annotation::annotation;
 pub(crate) use crate::parser::node_from_to;
 pub(crate) use crate::parser::occurrence::{
     individual_usage, occurrence_def, occurrence_usage, snapshot_usage, then_timeslice_usage,